Progress bars: indicatif
Extract upload out into a separate module since it's low-level logic

By the same token, it shouldn't really be bound to display, the channel
should be passed up the stack and delivered to a display module instead.
This can be a future improvement.

Extra colours on messaging too, and explicitly do nothing if there's no
syncing to do.