Use typed-path to make unix paths
Due to platform-dependent implementations of std Path and PathBuf, move
operations (and probably wipe-temp operations) were failing when the app
was running in windows and against a unix remote host -- which is what
it's designed to do.

Make sure remote paths use typed-path's Utf8UnixPath / Utf8UnixPathBuf
types instead, so that they'll be rendered correctly when creating
commands.